isLocal()

+++ database/migrations/2024_02_23_180755_add_collumn_sticker_id_to_product_table.php
@@ -21,9 +21,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropForeign(['sticker_id']);
-            $table->dropColumn(['sticker_id']);
-        });
+        if (app()->isLocal()) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropForeign(['sticker_id']);
+                $table->dropColumn(['sticker_id']);
+            });
+        }
     }
 };
