+++ database/migrations/2024_02_25_133248_create_images_table.php
@@ -25,6 +25,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('images');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('images');
+        }
     }
 };
